requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
Flask-SocketIO==5.3.6
python-socketio==5.10.0
eventlet==0.33.3
paho-mqtt==1.6.1
requests==2.31.0

# For LLM integration if using OpenAI API
openai==1.30.1
